During end2end-tests, feed input line by line with delay
Piping input into a cool-program too quickly may cause issues, including
null-dereferences. This is not a problem during normal, manual use of
such a program, but breaks automated tests. Thus, we ``simulate'' manual
input by giving the input to the program line by line, with a delay of
200ms in-between two lines.

This is one of the reasons for Issue #4, but does not completely solve
it yet.